requirements.txt
# general  modules
django >3.2,<3.3
python-decouple
Pillow
djangorestframework

# deployment module
gunicorn

# third party module
djoser
drf-yasg[validation]